chore: checkstyle and spotless
