+++ fabric-holder-component-api-v1/src/main/java/net/fabricmc/fabric/impl/holder/component/DataHolderComponentInitializer.java
@@ -1,13 +1,30 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package net.fabricmc.fabric.impl.holder.component;
+
+import java.io.Reader;
+import java.util.List;
+import java.util.Map;
 
 import com.google.gson.JsonElement;
 import com.mojang.logging.LogUtils;
-import com.mojang.serialization.Codec;
 import com.mojang.serialization.JsonOps;
+import org.slf4j.Logger;
 
-import net.fabricmc.fabric.api.holder.component.FabricDataComponentInitializer;
-
-import net.minecraft.core.Holder;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.Registry;
 import net.minecraft.core.component.DataComponentMap;
@@ -21,12 +38,7 @@ import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.packs.resources.Resource;
 import net.minecraft.util.StrictJsonParser;
 
-import org.slf4j.Logger;
-
-import java.io.Reader;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import net.fabricmc.fabric.api.holder.component.v1.FabricDataComponentInitializer;
 
 public class DataHolderComponentInitializer implements FabricDataComponentInitializer {
 	private static final Logger LOGGER = LogUtils.getLogger();
@@ -68,14 +80,15 @@ public class DataHolderComponentInitializer implements FabricDataComponentInitia
 			Identifier location,
 			Identifier id
 	) {
-
 		for (Resource resource : resources) {
 			try (Reader reader = resource.openAsReader()) {
 				JsonElement element = StrictJsonParser.parse(reader);
+
+				// TODO: Implement replace, also add required
 				DataHolderComponentFile<T, C> file = DataHolderComponentFile.codec(registryKey, componentType).parse(ops, element).getOrThrow();
 
 				for (Map.Entry<ResourceKey<T>, C> entry : file.components().entrySet()) {
-					var builder = context.builder(entry.getKey());
+					DataComponentMap.Builder builder = context.builder(entry.getKey());
 					builder.set(componentType, entry.getValue());
 				}
 			} catch (Exception e) {
