+++ fabric-holder-component-api-v1/src/main/java/net/fabricmc/fabric/mixin/holder/component/ReloadableServerResourcesMixin.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package net.fabricmc.fabric.mixin.holder.component;
 
 import java.util.concurrent.CompletableFuture;
@@ -7,14 +23,13 @@ import java.util.function.Supplier;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.llamalad7.mixinextras.sugar.Local;
-
-import net.fabricmc.fabric.impl.holder.component.FabricDataComponentInitializersImpl;
-
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
 import net.minecraft.server.ReloadableServerResources;
 import net.minecraft.server.packs.resources.ResourceManager;
+
+import net.fabricmc.fabric.impl.holder.component.FabricDataComponentInitializersImpl;
 
 @Mixin(ReloadableServerResources.class)
 public class ReloadableServerResourcesMixin {
@@ -23,7 +38,7 @@ public class ReloadableServerResourcesMixin {
 			Supplier<?> supplier,
 			Executor executor,
 			Operation<CompletableFuture<?>> original,
-			@Local(argsOnly = true, name = "resourceManager") ResourceManager resourceManager
+			@Local(argsOnly = true) ResourceManager resourceManager
 	) {
 		Supplier<?> wrappedSupplier = () -> ScopedValue.where(
 				FabricDataComponentInitializersImpl.RESOURCE_MANAGER,
